change flow a little bit
