--- packages/gatsby-theme-vtex/src/sdk/auth/OAuth/types.ts
@@ -1,1 +1,0 @@
-export type OAuthProvider = 'Google' | 'Facebook'
